Move files to utils
