+++ layoutlm/deprecated/examples/seq_labeling/utils/__init__.py
@@ -1,0 +1,1 @@
+# -*- coding: utf-8 -*-
